--- src/tests/api/v0/routers/hashes/mocks/__init__.py
@@ -1,5 +1,0 @@
-__all__ = [
-    'HashingServiceMock',
-]
-
-from tests.api.v0.routers.hashes.mocks.hashing_service_mock import HashingServiceMock
