--- src/tests/entities/hashing_algorithms/conftest.py
@@ -1,23 +1,0 @@
-import pytest
-
-from entities.hashing_algorithms import (
-    IHashingAlgorithm,
-    MD5HashingAlgorithm,
-    SHA1HashingAlgorithm,
-    SHA256HashingAlgorithm,
-)
-
-
-@pytest.fixture
-def md5() -> IHashingAlgorithm:
-    return MD5HashingAlgorithm()
-
-
-@pytest.fixture
-def sha1() -> IHashingAlgorithm:
-    return SHA1HashingAlgorithm()
-
-
-@pytest.fixture
-def sha256() -> IHashingAlgorithm:
-    return SHA256HashingAlgorithm()
