+++ src/tests/integration/api/v0/routers/hashes/mocks/__init__.py
@@ -1,0 +1,5 @@
+__all__ = [
+    'HashingServiceMock',
+]
+
+from tests.integration.api.v0.routers.hashes.mocks.hashing_service_mock import HashingServiceMock
