--- src/tests/services/hashing/factories/mocks/__init__.py
@@ -1,7 +1,0 @@
-__all__ = [
-    'MD5HashingAlgorithmMock',
-    'SHA1HashingAlgorithmMock',
-]
-
-from tests.services.hashing.factories.mocks.md5_mock import MD5HashingAlgorithmMock
-from tests.services.hashing.factories.mocks.sha1_mock import SHA1HashingAlgorithmMock
