--- src/tests/services/hashing/service/conftest.py
@@ -1,10 +1,0 @@
-import pytest
-
-from services.hashing import IHashingService, HashingService
-from tests.services.hashing.service.mocks import HashingAlgorithmFactoryMock
-
-
-@pytest.fixture
-def service() -> IHashingService:
-    factory_mock = HashingAlgorithmFactoryMock()
-    return HashingService(factory_mock)
