--- src/tests/services/hashing/service/mocks/__init__.py
@@ -1,9 +1,0 @@
-__all__ = [
-    'HashingAlgorithmFactoryMock',
-    'MD5HashingAlgorithmMock',
-    'SHA1HashingAlgorithmMock',
-]
-
-from tests.services.hashing.service.mocks.md5_mock import MD5HashingAlgorithmMock
-from tests.services.hashing.service.mocks.sha1_mock import SHA1HashingAlgorithmMock
-from tests.services.hashing.service.mocks.factory_mock import HashingAlgorithmFactoryMock
