--- src/tests/services/hashing/service/mocks/md5_mock.py
@@ -1,12 +1,0 @@
-__all__ = ['MD5HashingAlgorithmMock']
-
-from entities.hashing_algorithms import IHashingAlgorithm
-
-
-class MD5HashingAlgorithmMock(IHashingAlgorithm):
-    name: str = "md5"
-    bits: int = 128
-    is_secure: bool = False
-
-    def hash(self, data: str) -> str:
-        return 'MD5 Hash'
