+++ src/tests/unit/services/hashing/factories/mocks/md5_mock.py
@@ -1,0 +1,8 @@
+__all__ = ['MD5HashingAlgorithmMock']
+
+from app.entities.hashing_algorithms import IHashingAlgorithm
+
+
+class MD5HashingAlgorithmMock(IHashingAlgorithm):
+    def hash(self, data: str) -> str:
+        return 'MD5 Hash'
