+++ src/tests/unit/services/hashing/factories/mocks/sha1_mock.py
@@ -1,0 +1,8 @@
+__all__ = ['SHA1HashingAlgorithmMock']
+
+from app.entities.hashing_algorithms import IHashingAlgorithm
+
+
+class SHA1HashingAlgorithmMock(IHashingAlgorithm):
+    def hash(self, data: str) -> str:
+        return 'SHA1 Hash'
