+++ src/tests/unit/services/hashing/service/conftest.py
@@ -1,0 +1,10 @@
+import pytest
+
+from app.services.hashing import IHashingService, HashingService
+from tests.unit.services.hashing.service.mocks import HashingAlgorithmFactoryMock
+
+
+@pytest.fixture
+def service() -> IHashingService:
+    factory_mock = HashingAlgorithmFactoryMock()
+    return HashingService(factory_mock)
