+++ src/tests/unit/services/hashing/service/mocks/sha1_mock.py
@@ -1,0 +1,12 @@
+__all__ = ['SHA1HashingAlgorithmMock']
+
+from app.entities.hashing_algorithms import IHashingAlgorithm
+
+
+class SHA1HashingAlgorithmMock(IHashingAlgorithm):
+    name: str = "sha1"
+    bits: int = 160
+    is_secure: bool = False
+
+    def hash(self, data: str) -> str:
+        return 'SHA1 Hash'
